feat: BaseCandidateViewController共通化
共有化した時に選択候補の共重表示も消してしまった
